Added DBList and DBRow classes.
Started on actually building the GUI program a little. I didn't check for
errors, really, so it might have some, but the ideas are good, I think.
Basically, you have a DBList, and that is full of DBRows. The DBList will
have a populate() method, or something like that, and will thus get the
rows from the database, give them to the DBRows, and they will put them
on the screen. At least, that's roughly how I think it'll work. I have some
work to do yet, and I hope to God I'm just stupid today, because this is bending
my mind.