Added support for saving chosen columns.
It makes a config file the first time you pick, and then it opens that every
subsequent time and uses those columns instead of just asking you. I need to
add a button to reset it so you can choose again and make it so you can access
it in library form and supply the columns programmatically.